+++ Cargo.toml
@@ -14,7 +14,6 @@ members = [
   "juniper_hyper",
   "juniper_iron",
   "juniper_rocket",
-  "juniper_rocket_async",
   "juniper_subscriptions",
   "juniper_graphql_ws",
   "juniper_warp",
