+++ Makefile.toml
@@ -12,7 +12,7 @@ CARGO_MAKE_CARGO_ALL_FEATURES = ""
 [tasks.release]
 condition = { env_set = [ "RELEASE_LEVEL" ] }
 workspace = false
-env = { "CARGO_MAKE_WORKSPACE_SKIP_MEMBERS" = "integration_tests/*;examples/*;juniper_benchmarks;juniper_rocket_async" }
+env = { "CARGO_MAKE_WORKSPACE_SKIP_MEMBERS" = "integration_tests/*;examples/*;juniper_benchmarks;" }
 run_task = { name = "release-INTERNAL", fork = true }
 
 
@@ -39,7 +39,7 @@ args = ["release", "--config", "${CARGO_MAKE_WORKING_DIRECTORY}/../_build/releas
 [tasks.release-dry-run]
 condition = { env_set = [ "RELEASE_LEVEL" ] }
 workspace = false
-env = { "CARGO_MAKE_WORKSPACE_SKIP_MEMBERS" = "integration_tests/*;examples/*;juniper_benchmarks;juniper_rocket_async" }
+env = { "CARGO_MAKE_WORKSPACE_SKIP_MEMBERS" = "integration_tests/*;examples/*;juniper_benchmarks;" }
 run_task = { name = "release-dry-run-INTERNAL", fork = true }
 
 [tasks.release-some-dry-run]
@@ -53,7 +53,7 @@ run_task = { name = "release-some-dry-run-INTERNAL", fork = true }
 [tasks.release-dry-run-INTERNAL]
 private = true
 condition = { env_set = [ "RELEASE_LEVEL" ] }
-env = { "CARGO_MAKE_WORKSPACE_SKIP_MEMBERS" = "integration_tests/*;examples/*;juniper_benchmarks;juniper_rocket_async" }
+env = { "CARGO_MAKE_WORKSPACE_SKIP_MEMBERS" = "integration_tests/*;examples/*;juniper_benchmarks;" }
 description = "Run `cargo-release --dry-run` for every crate"
 command = "cargo-release"
 args = ["release", "--config", "${CARGO_MAKE_WORKING_DIRECTORY}/../_build/release.toml", "--dry-run", "${RELEASE_LEVEL}"]
@@ -68,7 +68,7 @@ args = ["release", "--config", "${CARGO_MAKE_WORKING_DIRECTORY}/../_build/releas
 [tasks.release-local-test]
 condition = { env_set = [ "RELEASE_LEVEL" ] }
 workspace = false
-env = { "CARGO_MAKE_WORKSPACE_SKIP_MEMBERS" = "integration_tests/*;examples/*;juniper_benchmarks;juniper_rocket_async" }
+env = { "CARGO_MAKE_WORKSPACE_SKIP_MEMBERS" = "integration_tests/*;examples/*;juniper_benchmarks;" }
 run_task = { name = "release-local-test-INTERNAL", fork = true }
