+++ juniper_rocket/Cargo.toml
@@ -12,8 +12,9 @@ documentation = "https://docs.rs/juniper_rocket"
 repository = "https://github.com/graphql-rust/juniper"
 
 [dependencies]
-juniper = { version = "0.15.6", path = "../juniper", default-features = false}
-rocket = { version = "0.4.10", default-features = false }
+futures = "0.3.1"
+juniper = { version = "0.15.6", path = "../juniper", default-features = false }
+rocket = { version = "0.5.0-rc.1", default-features = false }
 serde_json = "1.0.2"
 
 [dev-dependencies]
