+++ juniper_rocket/src/lib.rs
@@ -36,23 +36,23 @@ Check the LICENSE file for details.
 
 */
 
-#![doc(html_root_url = "https://docs.rs/juniper_rocket/0.2.0")]
-#![feature(decl_macro, proc_macro_hygiene)]
+#![doc(html_root_url = "https://docs.rs/juniper_rocket/0.7.1")]
 
-use std::io::{Cursor, Read};
+use std::{borrow::Cow, io::Cursor};
+
+use rocket::{
+    data::{self, FromData, ToByteUnit},
+    form::{error::ErrorKind, DataField, Error, Errors, FromForm, Options, ValueField},
+    http::{ContentType, Status},
+    outcome::Outcome::{Failure, Forward, Success},
+    response::{self, content, Responder, Response},
+    Data, Request,
+};
 
 use juniper::{
     http::{self, GraphQLBatchRequest},
-    DefaultScalarValue, FieldError, GraphQLType, InputValue, RootNode, ScalarValue,
-};
-use rocket::{
-    data::{FromDataSimple, Outcome as FromDataOutcome},
-    http::{ContentType, RawStr, Status},
-    request::{FormItems, FromForm, FromFormValue},
-    response::{content, Responder, Response},
-    Data,
-    Outcome::{Forward, Success},
-    Request,
+    DefaultScalarValue, FieldError, GraphQLSubscriptionType, GraphQLType, GraphQLTypeAsync,
+    InputValue, RootNode, ScalarValue,
 };
 
 /// Simple wrapper around an incoming GraphQL request
@@ -71,22 +71,22 @@ pub struct GraphQLResponse(pub Status, pub String);
 /// Generate an HTML page containing GraphiQL
 pub fn graphiql_source(
     graphql_endpoint_url: &str,
-    subscriptions_endpoint: Option<&str>,
+    subscriptions_endpoint_url: Option<&str>,
 ) -> content::Html<String> {
     content::Html(juniper::http::graphiql::graphiql_source(
         graphql_endpoint_url,
-        subscriptions_endpoint,
+        subscriptions_endpoint_url,
     ))
 }
 
 /// Generate an HTML page containing GraphQL Playground
 pub fn playground_source(
     graphql_endpoint_url: &str,
-    subscriptions_endpoint: Option<&str>,
+    subscriptions_endpoint_url: Option<&str>,
 ) -> content::Html<String> {
     content::Html(juniper::http::playground::playground_source(
         graphql_endpoint_url,
-        subscriptions_endpoint,
+        subscriptions_endpoint_url,
     ))
 }
 
@@ -94,7 +94,7 @@ impl<S> GraphQLRequest<S>
 where
     S: ScalarValue,
 {
-    /// Execute an incoming GraphQL query
+    /// Synchronously execute an incoming GraphQL query.
     pub fn execute_sync<CtxT, QueryT, MutationT, SubscriptionT>(
         &self,
         root_node: &RootNode<QueryT, MutationT, SubscriptionT, S>,
@@ -106,6 +106,33 @@ where
         SubscriptionT: GraphQLType<S, Context = CtxT>,
     {
         let response = self.0.execute_sync(root_node, context);
+        let status = if response.is_ok() {
+            Status::Ok
+        } else {
+            Status::BadRequest
+        };
+        let json = serde_json::to_string(&response).unwrap();
+
+        GraphQLResponse(status, json)
+    }
+
+    /// Asynchronously execute an incoming GraphQL query.
+    pub async fn execute<CtxT, QueryT, MutationT, SubscriptionT>(
+        &self,
+        root_node: &RootNode<'_, QueryT, MutationT, SubscriptionT, S>,
+        context: &CtxT,
+    ) -> GraphQLResponse
+    where
+        QueryT: GraphQLTypeAsync<S, Context = CtxT>,
+        QueryT::TypeInfo: Sync,
+        MutationT: GraphQLTypeAsync<S, Context = CtxT>,
+        MutationT::TypeInfo: Sync,
+        SubscriptionT: GraphQLSubscriptionType<S, Context = CtxT>,
+        SubscriptionT::TypeInfo: Sync,
+        CtxT: Sync,
+        S: Send + Sync,
+    {
+        let response = self.0.execute(root_node, context).await;
         let status = if response.is_ok() {
             Status::Ok
         } else {
@@ -130,10 +157,8 @@ impl GraphQLResponse {
     /// # Examples
     ///
     /// ```
-    /// # #![feature(decl_macro, proc_macro_hygiene)]
-    /// #
-    /// # use rocket::http::Cookies;
-    /// # use rocket::request::Form;
+    /// # use rocket::http::CookieJar;
+    /// # use rocket::form::Form;
     /// # use rocket::response::content;
     /// # use rocket::State;
     /// #
@@ -144,17 +169,17 @@ impl GraphQLResponse {
     /// #
     /// #[rocket::get("/graphql?<request..>")]
     /// fn get_graphql_handler(
-    ///     mut cookies: Cookies,
-    ///     context: State<Database>,
-    ///     request: Form<juniper_rocket::GraphQLRequest>,
-    ///     schema: State<Schema>,
+    ///     cookies: &CookieJar,
+    ///     context: &State<Database>,
+    ///     request: juniper_rocket::GraphQLRequest,
+    ///     schema: &State<Schema>,
     /// ) -> juniper_rocket::GraphQLResponse {
     ///     if cookies.get("user_id").is_none() {
     ///         let err = FieldError::new("User is not logged in", Value::null());
     ///         return juniper_rocket::GraphQLResponse::error(err);
     ///     }
     ///
-    ///     request.execute_sync(&schema, &context)
+    ///     request.execute_sync(&*schema, &*context)
     /// }
     /// ```
     pub fn error(error: FieldError) -> Self {
@@ -174,128 +199,173 @@ impl GraphQLResponse {
     }
 }
 
-impl<'f, S> FromForm<'f> for GraphQLRequest<S>
-where
-    S: ScalarValue,
-{
-    type Error = String;
+pub struct GraphQLContext<'f, S: ScalarValue> {
+    opts: Options,
+    query: Option<String>,
+    operation_name: Option<String>,
+    variables: Option<InputValue<S>>,
+    errors: Errors<'f>,
+}
 
-    fn from_form(form_items: &mut FormItems<'f>, strict: bool) -> Result<Self, String> {
-        let mut query = None;
-        let mut operation_name = None;
-        let mut variables = None;
+impl<'f, S: ScalarValue> GraphQLContext<'f, S> {
+    fn query(&mut self, value: String) {
+        if self.query.is_some() {
+            let error = Error::from(ErrorKind::Duplicate).with_name("query");
 
-        for form_item in form_items {
-            let (key, value) = form_item.key_value();
-            // Note: we explicitly decode in the match arms to save work rather
-            // than decoding every form item blindly.
-            match key.as_str() {
-                "query" => {
-                    if query.is_some() {
-                        return Err("Query parameter must not occur more than once".to_owned());
-                    } else {
-                        match value.url_decode() {
-                            Ok(v) => query = Some(v),
-                            Err(e) => return Err(e.to_string()),
-                        }
-                    }
-                }
-                "operation_name" => {
-                    if operation_name.is_some() {
-                        return Err(
-                            "Operation name parameter must not occur more than once".to_owned()
-                        );
-                    } else {
-                        match value.url_decode() {
-                            Ok(v) => operation_name = Some(v),
-                            Err(e) => return Err(e.to_string()),
-                        }
-                    }
-                }
-                "variables" => {
-                    if variables.is_some() {
-                        return Err("Variables parameter must not occur more than once".to_owned());
-                    } else {
-                        let decoded;
-                        match value.url_decode() {
-                            Ok(v) => decoded = v,
-                            Err(e) => return Err(e.to_string()),
-                        }
-                        variables = Some(
-                            serde_json::from_str::<InputValue<_>>(&decoded)
-                                .map_err(|err| err.to_string())?,
-                        );
-                    }
-                }
-                _ => {
-                    if strict {
-                        return Err(format!("Prohibited extra field '{}'", key));
-                    }
+            self.errors.push(error)
+        } else {
+            println!("{}", value);
+            self.query = Some(value);
+        }
+    }
+
+    fn operation_name(&mut self, value: String) {
+        if self.operation_name.is_some() {
+            let error = Error::from(ErrorKind::Duplicate).with_name("operation_name");
+
+            self.errors.push(error)
+        } else {
+            self.operation_name = Some(value);
+        }
+    }
+
+    fn variables(&mut self, value: String) {
+        if self.variables.is_some() {
+            let error = Error::from(ErrorKind::Duplicate).with_name("variables");
+
+            self.errors.push(error)
+        } else {
+            let parse_result = serde_json::from_str::<InputValue<S>>(&value);
+
+            match parse_result {
+                Ok(variables) => self.variables = Some(variables),
+                Err(e) => {
+                    let error = Error::from(ErrorKind::Validation(Cow::Owned(e.to_string())))
+                        .with_name("variables");
+
+                    self.errors.push(error);
                 }
             }
         }
+    }
+}
 
-        if let Some(query) = query {
-            Ok(GraphQLRequest(GraphQLBatchRequest::Single(
-                http::GraphQLRequest::new(query, operation_name, variables),
-            )))
-        } else {
-            Err("Query parameter missing".to_owned())
+#[rocket::async_trait]
+impl<'f, S> FromForm<'f> for GraphQLRequest<S>
+where
+    S: ScalarValue + Send,
+{
+    type Context = GraphQLContext<'f, S>;
+
+    fn init(opts: Options) -> Self::Context {
+        GraphQLContext {
+            opts,
+            query: None,
+            operation_name: None,
+            variables: None,
+            errors: Errors::new(),
+        }
+    }
+
+    fn push_value(ctx: &mut Self::Context, field: ValueField<'f>) {
+        match field.name.key().map(|key| key.as_str()) {
+            Some("query") => ctx.query(field.value.to_owned()),
+            Some("operation_name") => ctx.operation_name(field.value.to_owned()),
+            Some("variables") => ctx.variables(field.value.to_owned()),
+            Some(key) => {
+                if ctx.opts.strict {
+                    let error = Error::from(ErrorKind::Unknown).with_name(key);
+
+                    ctx.errors.push(error)
+                }
+            }
+            None => {
+                if ctx.opts.strict {
+                    let error = Error::from(ErrorKind::Unexpected);
+
+                    ctx.errors.push(error)
+                }
+            }
+        }
+    }
+
+    async fn push_data(ctx: &mut Self::Context, field: DataField<'f, '_>) {
+        if ctx.opts.strict {
+            let error = Error::from(ErrorKind::Unexpected).with_name(field.name);
+
+            ctx.errors.push(error)
+        }
+    }
+
+    fn finalize(mut ctx: Self::Context) -> rocket::form::Result<'f, Self> {
+        if ctx.query.is_none() {
+            let error = Error::from(ErrorKind::Missing).with_name("query");
+
+            ctx.errors.push(error)
+        }
+
+        match ctx.errors.is_empty() {
+            true => Ok(GraphQLRequest(GraphQLBatchRequest::Single(
+                http::GraphQLRequest::new(ctx.query.unwrap(), ctx.operation_name, ctx.variables),
+            ))),
+            false => Err(ctx.errors),
         }
     }
 }
 
-impl<'v, S> FromFormValue<'v> for GraphQLRequest<S>
+const BODY_LIMIT: u64 = 1024 * 100;
+
+#[rocket::async_trait]
+impl<'r, S> FromData<'r> for GraphQLRequest<S>
 where
     S: ScalarValue,
 {
     type Error = String;
 
-    fn from_form_value(form_value: &'v RawStr) -> Result<Self, Self::Error> {
-        let mut form_items = FormItems::from(form_value);
+    async fn from_data(
+        req: &'r Request<'_>,
+        data: Data<'r>,
+    ) -> data::Outcome<'r, Self, Self::Error> {
+        use rocket::tokio::io::AsyncReadExt as _;
 
-        Self::from_form(&mut form_items, true)
-    }
-}
-
-impl<S> FromDataSimple for GraphQLRequest<S>
-where
-    S: ScalarValue,
-{
-    type Error = String;
-
-    fn from_data(req: &Request, data: Data) -> FromDataOutcome<Self, Self::Error> {
         let content_type = req
             .content_type()
             .map(|ct| (ct.top().as_str(), ct.sub().as_str()));
         let is_json = match content_type {
             Some(("application", "json")) => true,
             Some(("application", "graphql")) => false,
-            _ => return Forward(data),
+            _ => return Box::pin(async move { Forward(data) }).await,
         };
 
-        let mut body = String::new();
-        data.open()
-            .read_to_string(&mut body)
-            .map_err(|e| (Status::InternalServerError, format!("{:?}", e)))?;
+        Box::pin(async move {
+            let mut body = String::new();
+            let mut reader = data.open(BODY_LIMIT.bytes());
+            if let Err(e) = reader.read_to_string(&mut body).await {
+                return Failure((Status::InternalServerError, format!("{:?}", e)));
+            }
 
-        Success(GraphQLRequest(if is_json {
-            serde_json::from_str(&body).map_err(|e| (Status::BadRequest, format!("{}", e)))?
-        } else {
-            GraphQLBatchRequest::Single(http::GraphQLRequest::new(body, None, None))
-        }))
+            Success(GraphQLRequest(if is_json {
+                match serde_json::from_str(&body) {
+                    Ok(req) => req,
+                    Err(e) => return Failure((Status::BadRequest, format!("{}", e))),
+                }
+            } else {
+                GraphQLBatchRequest::Single(http::GraphQLRequest::new(body, None, None))
+            }))
+        })
+        .await
     }
 }
 
-impl<'r> Responder<'r> for GraphQLResponse {
-    fn respond_to(self, _: &Request) -> Result<Response<'r>, Status> {
+impl<'r, 'o: 'r> Responder<'r, 'o> for GraphQLResponse {
+    fn respond_to(self, _req: &'r Request<'_>) -> response::Result<'o> {
         let GraphQLResponse(status, body) = self;
 
-        Ok(Response::build()
+        Response::build()
             .header(ContentType::new("application", "json"))
             .status(status)
-            .sized_body(Cursor::new(body))
-            .finalize())
+            .sized_body(body.len(), Cursor::new(body))
+            .ok()
     }
 }
 
@@ -303,40 +373,66 @@ impl<'r> Responder<'r> for GraphQLResponse {
 mod fromform_tests {
     use super::*;
     use juniper::InputValue;
-    use rocket::request::{FormItems, FromForm};
-    use std::str;
+    use rocket::{
+        form::{error::ErrorKind, Error, Form, Strict},
+        http::RawStr,
+    };
+    use std::borrow::Cow;
 
-    fn check_error(input: &str, error: &str, strict: bool) {
-        let mut items = FormItems::from(input);
-        let result: Result<GraphQLRequest, _> = GraphQLRequest::from_form(&mut items, strict);
-        assert!(result.is_err());
-        assert_eq!(result.unwrap_err(), error);
+    fn check_error(input: &str, expected_errors: Vec<Error>, strict: bool) {
+        let errors = if strict {
+            let result = Form::<Strict<GraphQLRequest>>::parse_encoded(RawStr::new(input));
+            assert!(result.is_err());
+            result.unwrap_err()
+        } else {
+            let result = Form::<GraphQLRequest>::parse_encoded(RawStr::new(input));
+            assert!(result.is_err());
+            result.unwrap_err()
+        };
+        assert_eq!(errors.len(), expected_errors.len());
+        for (error, expected) in errors.iter().zip(&expected_errors) {
+            match (&error.kind, &expected.kind) {
+                (ErrorKind::Unknown, ErrorKind::Unknown) => (),
+                (kind_a, kind_b) => assert_eq!(kind_a, kind_b),
+            };
+            assert_eq!(error.name, expected.name);
+            assert_eq!(error.value, expected.value);
+            assert_eq!(error.entity, expected.entity);
+        }
     }
 
     #[test]
     fn test_empty_form() {
-        check_error("", "Query parameter missing", false);
+        check_error(
+            "",
+            vec![Error::from(ErrorKind::Missing).with_name("query")],
+            false,
+        );
     }
 
     #[test]
     fn test_no_query() {
         check_error(
             "operation_name=foo&variables={}",
-            "Query parameter missing",
+            vec![Error::from(ErrorKind::Missing).with_name("query")],
             false,
         );
     }
 
     #[test]
     fn test_strict() {
-        check_error("query=test&foo=bar", "Prohibited extra field \'foo\'", true);
+        check_error(
+            "query=test&foo=bar",
+            vec![Error::from(ErrorKind::Unknown).with_name("foo")],
+            true,
+        );
     }
 
     #[test]
     fn test_duplicate_query() {
         check_error(
             "query=foo&query=bar",
-            "Query parameter must not occur more than once",
+            vec![Error::from(ErrorKind::Duplicate).with_name("query")],
             false,
         );
     }
@@ -345,7 +441,7 @@ mod fromform_tests {
     fn test_duplicate_operation_name() {
         check_error(
             "query=test&operation_name=op1&operation_name=op2",
-            "Operation name parameter must not occur more than once",
+            vec![Error::from(ErrorKind::Duplicate).with_name("operation_name")],
             false,
         );
     }
@@ -354,7 +450,7 @@ mod fromform_tests {
     fn test_duplicate_variables() {
         check_error(
             "query=test&variables={}&variables={}",
-            "Variables parameter must not occur more than once",
+            vec![Error::from(ErrorKind::Duplicate).with_name("variables")],
             false,
         );
     }
@@ -363,16 +459,18 @@ mod fromform_tests {
     fn test_variables_invalid_json() {
         check_error(
             "query=test&variables=NOT_JSON",
-            "expected value at line 1 column 1",
+            vec![Error::from(ErrorKind::Validation(Cow::Owned(
+                "expected value at line 1 column 1".to_owned(),
+            )))
+            .with_name("variables")],
             false,
         );
     }
 
     #[test]
     fn test_variables_valid_json() {
-        let form_string = r#"query=test&variables={"foo":"bar"}"#;
-        let mut items = FormItems::from(form_string);
-        let result = GraphQLRequest::from_form(&mut items, false);
+        let result: Result<GraphQLRequest, Errors> =
+            Form::parse_encoded(RawStr::new(r#"query=test&variables={"foo":"bar"}"#));
         assert!(result.is_ok());
         let variables = ::serde_json::from_str::<InputValue>(r#"{"foo":"bar"}"#).unwrap();
         let expected = GraphQLRequest(GraphQLBatchRequest::Single(http::GraphQLRequest::new(
@@ -385,10 +483,9 @@ mod fromform_tests {
 
     #[test]
     fn test_variables_encoded_json() {
-        let form_string = r#"query=test&variables={"foo": "x%20y%26%3F+z"}"#;
-        let mut items = FormItems::from(form_string);
-        let result = GraphQLRequest::from_form(&mut items, false);
-        assert!(result.is_ok());
+        let result: Result<GraphQLRequest, Errors> = Form::parse_encoded(RawStr::new(
+            r#"query=test&variables={"foo":"x%20y%26%3F+z"}"#,
+        ));
         let variables = ::serde_json::from_str::<InputValue>(r#"{"foo":"x y&? z"}"#).unwrap();
         let expected = GraphQLRequest(GraphQLBatchRequest::Single(http::GraphQLRequest::new(
             "test".to_string(),
@@ -400,9 +497,9 @@ mod fromform_tests {
 
     #[test]
     fn test_url_decode() {
-        let form_string = "query=%25foo%20bar+baz%26%3F&operation_name=test";
-        let mut items = FormItems::from(form_string);
-        let result: Result<GraphQLRequest, _> = GraphQLRequest::from_form(&mut items, false);
+        let result: Result<GraphQLRequest, Errors> = Form::parse_encoded(RawStr::new(
+            "query=%25foo%20bar+baz%26%3F&operation_name=test",
+        ));
         assert!(result.is_ok());
         let expected = GraphQLRequest(GraphQLBatchRequest::Single(http::GraphQLRequest::new(
             "%foo bar baz&?".to_string(),
@@ -415,6 +512,9 @@ mod fromform_tests {
 
 #[cfg(test)]
 mod tests {
+
+    use futures;
+
     use juniper::{
         http::tests as http_tests,
         tests::fixtures::starwars::schema::{Database, Query},
@@ -423,30 +523,28 @@ mod tests {
     use rocket::{
         self, get,
         http::ContentType,
-        local::{Client, LocalRequest},
-        post,
-        request::Form,
-        routes, Rocket, State,
+        local::asynchronous::{Client, LocalResponse},
+        post, routes, Build, Rocket, State,
     };
 
     type Schema = RootNode<'static, Query, EmptyMutation<Database>, EmptySubscription<Database>>;
 
     #[get("/?<request..>")]
     fn get_graphql_handler(
-        context: State<Database>,
-        request: Form<super::GraphQLRequest>,
-        schema: State<Schema>,
+        context: &State<Database>,
+        request: super::GraphQLRequest,
+        schema: &State<Schema>,
     ) -> super::GraphQLResponse {
-        request.execute_sync(&schema, &context)
+        request.execute_sync(&*schema, &*context)
     }
 
     #[post("/", data = "<request>")]
     fn post_graphql_handler(
-        context: State<Database>,
+        context: &State<Database>,
         request: super::GraphQLRequest,
-        schema: State<Schema>,
+        schema: &State<Schema>,
     ) -> super::GraphQLResponse {
-        request.execute_sync(&schema, &context)
+        request.execute_sync(&*schema, &*context)
     }
 
     struct TestRocketIntegration {
@@ -455,86 +553,90 @@ mod tests {
 
     impl http_tests::HttpIntegration for TestRocketIntegration {
         fn get(&self, url: &str) -> http_tests::TestResponse {
-            let req = &self.client.get(url);
-            make_test_response(req)
+            let req = self.client.get(url);
+            let req = futures::executor::block_on(req.dispatch());
+            futures::executor::block_on(make_test_response(req))
         }
 
         fn post_json(&self, url: &str, body: &str) -> http_tests::TestResponse {
-            let req = &self.client.post(url).header(ContentType::JSON).body(body);
-            make_test_response(req)
+            let req = self.client.post(url).header(ContentType::JSON).body(body);
+            let req = futures::executor::block_on(req.dispatch());
+            futures::executor::block_on(make_test_response(req))
         }
 
         fn post_graphql(&self, url: &str, body: &str) -> http_tests::TestResponse {
-            let req = &self
+            let req = self
                 .client
                 .post(url)
                 .header(ContentType::new("application", "graphql"))
                 .body(body);
-            make_test_response(req)
+            let req = futures::executor::block_on(req.dispatch());
+            futures::executor::block_on(make_test_response(req))
         }
     }
 
-    #[test]
-    fn test_rocket_integration() {
+    #[rocket::async_test]
+    async fn test_rocket_integration() {
         let rocket = make_rocket();
-        let client = Client::new(rocket).expect("valid rocket");
+        let client = Client::untracked(rocket).await.expect("valid rocket");
         let integration = TestRocketIntegration { client };
 
         http_tests::run_http_test_suite(&integration);
     }
 
-    #[test]
-    fn test_operation_names() {
+    #[rocket::async_test]
+    async fn test_operation_names() {
         #[post("/", data = "<request>")]
         fn post_graphql_assert_operation_name_handler(
-            context: State<Database>,
+            context: &State<Database>,
             request: super::GraphQLRequest,
-            schema: State<Schema>,
+            schema: &State<Schema>,
         ) -> super::GraphQLResponse {
             assert_eq!(request.operation_names(), vec![Some("TestQuery")]);
-            request.execute_sync(&schema, &context)
+            request.execute_sync(&*schema, &*context)
         }
 
         let rocket = make_rocket_without_routes()
             .mount("/", routes![post_graphql_assert_operation_name_handler]);
-        let client = Client::new(rocket).expect("valid rocket");
+        let client = Client::untracked(rocket).await.expect("valid rocket");
 
-        let req = client
+        let resp = client
             .post("/")
             .header(ContentType::JSON)
-            .body(r#"{"query": "query TestQuery {hero{name}}", "operationName": "TestQuery"}"#);
-        let resp = make_test_response(&req);
+            .body(r#"{"query": "query TestQuery {hero{name}}", "operationName": "TestQuery"}"#)
+            .dispatch()
+            .await;
+        let resp = make_test_response(resp);
 
-        assert_eq!(resp.status_code, 200);
+        assert_eq!(resp.await.status_code, 200);
     }
 
-    fn make_rocket() -> Rocket {
+    fn make_rocket() -> Rocket<Build> {
         make_rocket_without_routes().mount("/", routes![post_graphql_handler, get_graphql_handler])
     }
 
-    fn make_rocket_without_routes() -> Rocket {
-        rocket::ignite().manage(Database::new()).manage(Schema::new(
+    fn make_rocket_without_routes() -> Rocket<Build> {
+        Rocket::build().manage(Database::new()).manage(Schema::new(
             Query,
             EmptyMutation::<Database>::new(),
             EmptySubscription::<Database>::new(),
         ))
     }
 
-    fn make_test_response(request: &LocalRequest) -> http_tests::TestResponse {
-        let mut response = request.clone().dispatch();
+    async fn make_test_response(response: LocalResponse<'_>) -> http_tests::TestResponse {
         let status_code = response.status().code as i32;
         let content_type = response
             .content_type()
             .expect("No content type header from handler")
             .to_string();
         let body = response
-            .body()
-            .expect("No body returned from GraphQL handler")
-            .into_string();
+            .into_string()
+            .await
+            .expect("No body returned from GraphQL handler");
 
         http_tests::TestResponse {
             status_code,
-            body,
+            body: Some(body),
             content_type,
         }
     }
