+++ juniper_rocket/tests/custom_response_tests.rs
@@ -1,6 +1,5 @@
-use rocket::http::Status;
-
 use juniper_rocket::GraphQLResponse;
+use rocket::http::Status;
 
 #[test]
 fn test_graphql_response_is_public() {
