--- juniper_rocket_async/Cargo.toml
@@ -1,21 +1,0 @@
-[package]
-name = "juniper_rocket_async"
-version = "0.5.1"
-edition = "2018"
-authors = [
-    "Magnus Hallin <mhallin@fastmail.com>",
-    "Christoph Herzog <chris@theduke.at>",
-]
-description = "Juniper GraphQL integration with Rocket"
-license = "BSD-2-Clause"
-documentation = "https://docs.rs/juniper_rocket"
-repository = "https://github.com/graphql-rust/juniper"
-
-[dependencies]
-futures = "0.3.1"
-juniper = { version = "0.15.6", path = "../juniper", default-features = false }
-rocket = { version = "0.5.0-rc.1", default-features = false }
-serde_json = "1.0.2"
-
-[dev-dependencies]
-juniper = { version = "0.15.6", path = "../juniper", features = ["expose-test-schema"] }
