--- juniper_rocket_async/README.md
@@ -1,35 +1,0 @@
-# juniper_rocket
-
-This repository contains the [Rocket][Rocket] web server integration for 
-[Juniper][Juniper], a [GraphQL][GraphQL] implementation for Rust.
-
-## Documentation
-
-For documentation, including guides and examples, check out [Juniper][Juniper].
-
-A basic usage example can also be found in the [Api documentation][documentation].
-
-## Examples
-
-Check [examples/rocket_server.rs][example] for example code of a working Rocket 
-server with GraphQL handlers.
-
-## Links
-
-* [Juniper][Juniper]
-* [Api Reference][documetation]
-* [Rocket][Iron]
-
-## License
-
-This project is under the BSD-2 license.
-
-Check the LICENSE file for details.
-
-[Rocket]: https://rocket.rs
-[Juniper]: https://github.com/graphql-rust/juniper
-[GraphQL]: http://graphql.org
-[documentation]: https://docs.rs/juniper_rocket
-[example]: https://github.com/graphql-rust/juniper/blob/master/juniper_rocket/examples/rocket_server.rs
-
-
