--- juniper_rocket_async/examples/rocket_server.rs
@@ -1,48 +1,0 @@
-use juniper::{
-    tests::fixtures::starwars::schema::{Database, Query},
-    EmptyMutation, EmptySubscription, RootNode,
-};
-use rocket::{response::content, Rocket, State};
-
-type Schema = RootNode<'static, Query, EmptyMutation<Database>, EmptySubscription<Database>>;
-
-#[rocket::get("/")]
-fn graphiql() -> content::Html<String> {
-    juniper_rocket_async::graphiql_source("/graphql", None)
-}
-
-#[rocket::get("/graphql?<request>")]
-fn get_graphql_handler(
-    context: &State<Database>,
-    request: juniper_rocket_async::GraphQLRequest,
-    schema: &State<Schema>,
-) -> juniper_rocket_async::GraphQLResponse {
-    request.execute_sync(&*schema, &*context)
-}
-
-#[rocket::post("/graphql", data = "<request>")]
-fn post_graphql_handler(
-    context: &State<Database>,
-    request: juniper_rocket_async::GraphQLRequest,
-    schema: &State<Schema>,
-) -> juniper_rocket_async::GraphQLResponse {
-    request.execute_sync(&*schema, &*context)
-}
-
-#[rocket::main]
-async fn main() {
-    Rocket::build()
-        .manage(Database::new())
-        .manage(Schema::new(
-            Query,
-            EmptyMutation::<Database>::new(),
-            EmptySubscription::<Database>::new(),
-        ))
-        .mount(
-            "/",
-            rocket::routes![graphiql, get_graphql_handler, post_graphql_handler],
-        )
-        .launch()
-        .await
-        .expect("server to launch");
-}
